split the JSONP and CORS tests to separate files
